src/ClassFramework.Pipelines.Tests/Entity/Features/SetRecordFeatureTests.cs
namespace ClassFramework.Pipelines.Tests.Entity.Features;

public class SetRecordFeatureTests : TestBase<Pipelines.Entity.Features.SetRecordFeature>
{
    public class Process : SetRecordFeatureTests
    {
        [Fact]
        public void Throws_On_Null_Context()
        {
            // Arrange
            var sut = CreateSut();

            // Act & Assert
            sut.Invoking(x => x.Process(context: null!))
               .Should().Throw<ArgumentNullException>().WithParameterName("context");
        }

        [Theory]
        [InlineData(false, false)]
        [InlineData(true, true)]
        public void Sets_Record_Based_On_Setting(bool createRecordSettingValue, bool expectedRecordValue)
        {
            // Arrange
            var sourceModel = CreateModel();
            InitializeParser();
            var sut = CreateSut();
            var model = new ClassBuilder();
            var settings = CreateSettingsForEntity(createRecord:  createRecordSettingValue);
            var context = new PipelineContext<IConcreteTypeBuilder, EntityContext>(model, new EntityContext(sourceModel, settings.Build(), CultureInfo.InvariantCulture));

            // Act
            var result = sut.Process(context);

            // Assert
            result.IsSuccessful().Should().BeTrue();
            model.Record.Should().Be(expectedRecordValue);
        }
    }
}
